requirements.txt
google-adk>=1.4.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.12.8
beautifulsoup4>=4.13.4
